break cycle dependency on LDAP feature
