+++ packages/cli/src/Ldap/helpers.ts
@@ -10,7 +10,6 @@ import type { Role } from '@db/entities/Role';
 import { User } from '@db/entities/User';
 import { AuthIdentity } from '@db/entities/AuthIdentity';
 import type { AuthProviderSyncHistory } from '@db/entities/AuthProviderSyncHistory';
-import { LdapManager } from './LdapManager.ee';
 
 import {
 	BINARY_AD_ATTRIBUTES,
@@ -20,9 +19,8 @@ import {
 	LDAP_LOGIN_LABEL,
 } from './constants';
 import type { ConnectionSecurity, LdapConfig } from './types';
-import { ApplicationError, jsonParse } from 'n8n-workflow';
+import { jsonParse } from 'n8n-workflow';
 import { License } from '@/License';
-import { InternalHooks } from '@/InternalHooks';
 import {
 	getCurrentAuthenticationMethod,
 	isEmailCurrentAuthenticationMethod,
@@ -30,12 +28,10 @@ import {
 	setCurrentAuthenticationMethod,
 } from '@/sso/ssoHelpers';
 import { RoleService } from '@/services/role.service';
-import { Logger } from '@/Logger';
 import { UserRepository } from '@db/repositories/user.repository';
 import { SettingsRepository } from '@db/repositories/settings.repository';
 import { AuthProviderSyncHistoryRepository } from '@db/repositories/authProviderSyncHistory.repository';
 import { AuthIdentityRepository } from '@db/repositories/authIdentity.repository';
-import { BadRequestError } from '@/errors/response-errors/bad-request.error';
 import { InternalServerError } from '@/errors/response-errors/internal-server.error';
 
 /**
@@ -137,7 +133,7 @@ export const setGlobalLdapConfigVariables = async (ldapConfig: LdapConfig): Prom
 	setLdapLoginLabel(ldapConfig.loginLabel);
 };
 
-const resolveEntryBinaryAttributes = (entry: LdapUser): LdapUser => {
+export const resolveEntryBinaryAttributes = (entry: LdapUser): LdapUser => {
 	Object.entries(entry)
 		.filter(([k]) => BINARY_AD_ATTRIBUTES.includes(k))
 		.forEach(([k]) => {
@@ -148,63 +144,6 @@ const resolveEntryBinaryAttributes = (entry: LdapUser): LdapUser => {
 
 export const resolveBinaryAttributes = (entries: LdapUser[]): void => {
 	entries.forEach((entry) => resolveEntryBinaryAttributes(entry));
-};
-
-/**
- * Update the LDAP configuration in the database
- */
-export const updateLdapConfig = async (ldapConfig: LdapConfig): Promise<void> => {
-	const { valid, message } = validateLdapConfigurationSchema(ldapConfig);
-
-	if (!valid) {
-		throw new ApplicationError(message);
-	}
-
-	if (ldapConfig.loginEnabled && getCurrentAuthenticationMethod() === 'saml') {
-		throw new BadRequestError('LDAP cannot be enabled if SSO in enabled');
-	}
-
-	LdapManager.updateConfig({ ...ldapConfig });
-
-	ldapConfig.bindingAdminPassword = Container.get(Cipher).encrypt(ldapConfig.bindingAdminPassword);
-
-	if (!ldapConfig.loginEnabled) {
-		ldapConfig.synchronizationEnabled = false;
-		const ldapUsers = await getLdapUsers();
-		if (ldapUsers.length) {
-			await deleteAllLdapIdentities();
-		}
-	}
-
-	await Container.get(SettingsRepository).update(
-		{ key: LDAP_FEATURE_NAME },
-		{ value: JSON.stringify(ldapConfig), loadOnStartup: true },
-	);
-	await setGlobalLdapConfigVariables(ldapConfig);
-};
-
-/**
- * Handle the LDAP initialization.
- * If it's the first run of this feature, all the default data is created in the database
- */
-export const handleLdapInit = async (): Promise<void> => {
-	if (!isLdapEnabled()) return;
-
-	const ldapConfig = await getLdapConfig();
-
-	try {
-		await setGlobalLdapConfigVariables(ldapConfig);
-	} catch (error) {
-		Container.get(Logger).warn(
-			`Cannot set LDAP login enabled state when an authentication method other than email or ldap is active (current: ${getCurrentAuthenticationMethod()})`,
-			// eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-			error,
-		);
-	}
-
-	// init LDAP manager with the current
-	// configuration
-	LdapManager.init(ldapConfig);
 };
 
 export const createFilter = (filter: string, userFilter: string) => {
@@ -218,69 +157,6 @@ export const createFilter = (filter: string, userFilter: string) => {
 export const escapeFilter = (filter: string): string => {
 	//@ts-ignore
 	return new Filter().escape(filter); /* eslint-disable-line */
-};
-
-/**
- * Find and authenticate user in the LDAP server.
- */
-export const findAndAuthenticateLdapUser = async (
-	loginId: string,
-	password: string,
-	loginIdAttribute: string,
-	userFilter: string,
-): Promise<LdapUser | undefined> => {
-	const ldapService = LdapManager.getInstance().service;
-
-	// Search for the user with the administrator binding using the
-	// the Login ID attribute and whatever was inputted in the UI's
-	// email input.
-	let searchResult: LdapUser[] = [];
-
-	try {
-		searchResult = await ldapService.searchWithAdminBinding(
-			createFilter(`(${loginIdAttribute}=${escapeFilter(loginId)})`, userFilter),
-		);
-	} catch (e) {
-		if (e instanceof Error) {
-			void Container.get(InternalHooks).onLdapLoginSyncFailed({
-				error: e.message,
-			});
-			Container.get(Logger).error('LDAP - Error during search', { message: e.message });
-		}
-		return undefined;
-	}
-
-	if (!searchResult.length) {
-		return undefined;
-	}
-
-	// In the unlikely scenario that more than one user is found (
-	// can happen depending on how the LDAP database is structured
-	// and the LDAP configuration), return the last one found as it
-	// should be the less important in the hierarchy.
-	let user = searchResult.pop();
-
-	if (user === undefined) {
-		user = { dn: '' };
-	}
-
-	try {
-		// Now with the user distinguished name (unique identifier
-		// for the user) and the password, attempt to validate the
-		// user by binding
-		await ldapService.validUser(user.dn, password);
-	} catch (e) {
-		if (e instanceof Error) {
-			Container.get(Logger).error('LDAP - Error validating user against LDAP server', {
-				message: e.message,
-			});
-		}
-		return undefined;
-	}
-
-	resolveEntryBinaryAttributes(user);
-
-	return user;
 };
 
 /**
@@ -475,6 +351,6 @@ export const updateLdapUserOnLocalDb = async (identity: AuthIdentity, data: Part
 	}
 };
 
-const deleteAllLdapIdentities = async () => {
+export const deleteAllLdapIdentities = async () => {
 	return Container.get(AuthIdentityRepository).delete({ providerType: 'ldap' });
 };
