+++ packages/cli/src/controllers/ldap.controller.ts
@@ -1,13 +1,26 @@
 import pick from 'lodash/pick';
 import { Authorized, Get, Post, Put, RestController, RequireGlobalScope } from '@/decorators';
-import { getLdapConfig, getLdapSynchronizations, updateLdapConfig } from '@/Ldap/helpers';
+import {
+	deleteAllLdapIdentities,
+	getLdapConfig,
+	getLdapSynchronizations,
+	getLdapUsers,
+	setGlobalLdapConfigVariables,
+	validateLdapConfigurationSchema,
+} from '@/Ldap/helpers';
 import { LdapManager } from '@/Ldap/LdapManager.ee';
 import type { LdapService } from '@/Ldap/LdapService.ee';
 import type { LdapSync } from '@/Ldap/LdapSync.ee';
+import type { LdapConfig } from '@/Ldap/types';
 import { LdapConfiguration } from '@/Ldap/types';
-import { NON_SENSIBLE_LDAP_CONFIG_PROPERTIES } from '@/Ldap/constants';
+import { LDAP_FEATURE_NAME, NON_SENSIBLE_LDAP_CONFIG_PROPERTIES } from '@/Ldap/constants';
 import { InternalHooks } from '@/InternalHooks';
 import { BadRequestError } from '@/errors/response-errors/bad-request.error';
+import { ApplicationError } from '../../../workflow/src';
+import { getCurrentAuthenticationMethod } from '@/sso/ssoHelpers';
+import Container from 'typedi';
+import { SettingsRepository } from '@/databases/repositories/settings.repository';
+import { Cipher } from 'n8n-core';
 
 @Authorized()
 @RestController('/ldap')
@@ -42,7 +55,7 @@ export class LdapController {
 	@RequireGlobalScope('ldap:manage')
 	async updateConfig(req: LdapConfiguration.Update) {
 		try {
-			await updateLdapConfig(req.body);
+			await this.updateLdapConfig(req.body);
 		} catch (error) {
 			throw new BadRequestError((error as { message: string }).message);
 		}
@@ -73,4 +86,36 @@ export class LdapController {
 			throw new BadRequestError((error as { message: string }).message);
 		}
 	}
+
+	private updateLdapConfig = async (ldapConfig: LdapConfig): Promise<void> => {
+		const { valid, message } = validateLdapConfigurationSchema(ldapConfig);
+
+		if (!valid) {
+			throw new ApplicationError(message);
+		}
+
+		if (ldapConfig.loginEnabled && getCurrentAuthenticationMethod() === 'saml') {
+			throw new BadRequestError('LDAP cannot be enabled if SSO in enabled');
+		}
+
+		LdapManager.updateConfig({ ...ldapConfig });
+
+		ldapConfig.bindingAdminPassword = Container.get(Cipher).encrypt(
+			ldapConfig.bindingAdminPassword,
+		);
+
+		if (!ldapConfig.loginEnabled) {
+			ldapConfig.synchronizationEnabled = false;
+			const ldapUsers = await getLdapUsers();
+			if (ldapUsers.length) {
+				await deleteAllLdapIdentities();
+			}
+		}
+
+		await Container.get(SettingsRepository).update(
+			{ key: LDAP_FEATURE_NAME },
+			{ value: JSON.stringify(ldapConfig), loadOnStartup: true },
+		);
+		await setGlobalLdapConfigVariables(ldapConfig);
+	};
 }
